Update bench script measurement collection
Measurements are now collected from stdout and stderr in addition to the
results file (which continues to work unchanged). Also, if a metric is
specified multiple times, its values are added automatically. All of
this is documented in more detail in the readme.

Log lines in the UI containing measurements are bolded.